requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
pytesseract>=0.3.10
pdfplumber>=0.7.0
PyYAML>=6.0
paramiko>=3.0.0
requests>=2.28.0
Pillow>=9.0.0
python-docx>=0.8.11
rarfile>=4.0 